requirement.txt
google-adk[extensions]
language-tool-python
requests
python-dotenv